+++ projects/ngx-location-picker/src/lib/types/cascading-rules.model.ts
@@ -1,0 +1,25 @@
+export interface CascadingRulesModel {
+  type?: CascadingRulesType;
+  addressOptions?: ReverseGeocodeOptionsModel;
+  locationOptions?: GeoFeaturesOptionsModel;
+}
+
+export enum CascadingRulesType {
+  LOCATION = 'location',
+  ADDRESS = 'address'
+}
+
+export interface ReverseGeocodeOptionsModel {
+  buffer?: number;
+  count?: number;
+  relevance?: boolean;
+}
+
+export interface GeoFeaturesOptionsModel {
+  mapService?: string;
+  layerIds?: string;
+  returnGeometry?: boolean;
+  tolerance?: number;
+  buffer?: number;
+  count?: number;
+}